Top screen = bottom screen
made bottom and top screen exactly the same it terms of  background, course grid, starting positions of players, restrictions on cameras, and player mobility values. In editor bottom screen is exactly 1,000 units above top screen. Only thing is the UI that splits the screen cuts into the top screen a bit so it looks slightly diffrent.